Setup the Order table and a feature of new table
